bitstream_io

Module huffman

Traits and implementations for reading or writing Huffman codes from or to a stream.

Structs

LimitedUnary

A limited unary reader which stops at the given maximum.

Traits

FromBits

A trait for building a final value from individual bits

ToBits

For building individual bits from a final value