Struct bitstream_io::huffman::WriteHuffmanTreeBE

pub struct WriteHuffmanTreeBE<T: Ord> { /* fields omitted */ }

Methods

impl<T: Ord + Copy> WriteHuffmanTreeBE<T>

fn get(&self, value: T) -> (u32, u64)

fn compile(values: &[(Vec<u8>, T)])
-> Result<WriteHuffmanTreeBE<T>, HuffmanTreeError>

Given a slice of bits/value pairs, compiles a Huffman tree for writing. Bits must be 0 or 1 and are always written to the stream from least-significant in the list to most signficant (which makes them easier to read for humans). Value can be anything that implements the Copy trait.

Example

use bitstream_io::huffman::ReadHuffmanTree;
assert!(ReadHuffmanTree::compile(&[(vec![0], 1i32),
                                   (vec![1, 0], 2i32),
                                   (vec![1, 1], 3i32)]).is_ok());