bitstream_io

Module read

Traits and implementations for reading bits from a stream.

Structs

BitReader

For reading non-aligned bits from a stream of bytes in a given endianness.

ByteReader

For reading aligned bytes from a stream of bytes in a given endianness.

Traits

BitRead

A trait for anything that can read a variable number of potentially un-aligned values from an input stream

BitRead2

A compatibility trait for older code implementing BitRead

ByteRead

A trait for anything that can read aligned values from an input stream

FromBitStream

Implemented by complex types that don’t require any additional context to parse themselves from a reader. Analogous to std::str::FromStr.

FromBitStreamUsing

Implemented by complex types that consume some immutable context to parse themselves from a reader.

FromBitStreamWith

Implemented by complex types that require some immutable context to parse themselves from a reader.

FromByteStream

Implemented by complex types that don’t require any additional context to parse themselves from a reader. Analagous to FromStr.

FromByteStreamUsing

Implemented by complex types that consume some additional context to parse themselves from a reader.

FromByteStreamWith

Implemented by complex types that require some additional context to parse themselves from a reader. Analagous to FromStr.