Struct bitstream_io::read::BitReaderBE

pub struct BitReaderBE<'a> { /* fields omitted */ }

A wrapper for reading values from a big-endian stream.

Methods

impl<'a> BitReaderBE<'a>

fn new(reader: &mut BufRead) -> BitReaderBE

Wraps a big-endian reader around a BufRead reference.

A BufRead is required because this reader is liable to make many small reads to the stream in normal operation, so reading from the buffer directly is preferable to making many calls to read_exact.