bitstream_io::read

Struct ByteReader

pub struct ByteReader<R: Read, E: Endianness> { /* private fields */ }

For reading aligned bytes from a stream of bytes in a given endianness.

This only reads aligned values and maintains no internal state.

Implementations

impl<R: Read, E: Endianness> ByteReader<R, E>

pub fn new(reader: R) -> ByteReader<R, E>

Wraps a ByteReader around something that implements Read

pub fn endian(reader: R, _endian: E) -> ByteReader<R, E>

Wraps a ByteReader around something that implements Read with the given endianness.

pub fn into_reader(self) -> R

Unwraps internal reader and disposes of ByteReader.

pub fn reader(&mut self) -> &mut R

Provides mutable reference to internal reader

pub fn into_bitreader(self) -> BitReader<R, E>

Converts ByteReader to BitReader in the same endianness.

pub fn bitreader(&mut self) -> BitReader<&mut R, E>

Provides temporary BitReader in the same endianness.

Warning

Any unread bits left over when BitReader is dropped are lost.