Struct bitstream_io::BitQueueLE

pub struct BitQueueLE<N: Numeric> { /* fields omitted */ }

A wrapper around some unsigned type to turn it into a little-endian queue.

Methods

impl<N: Numeric> BitQueueLE<N>

fn from_value(value: N, bits: u32) -> BitQueueLE<N>

Wraps an existing value in a little-endian bit queue with the given number of bits. It is up to the programmer to ensure that the value is not larger than the given number of bits can contain.

fn new() -> BitQueueLE<N>

Creates an empty queue with a value and bit count of 0.