bitstream_io

Trait Integer

pub trait Integer {
    // Required methods
    fn read<const BITS: u32, R: BitRead + ?Sized>(
        reader: &mut R,
    ) -> Result<Self>
       where Self: Sized;
    fn read_var<const MAX: u32, R>(
        reader: &mut R,
        bits: BitCount<MAX>,
    ) -> Result<Self>
       where R: BitRead + ?Sized,
             Self: Sized;
    fn write<const BITS: u32, W: BitWrite + ?Sized>(
        self,
        writer: &mut W,
    ) -> Result<()>;
    fn write_var<const MAX: u32, W: BitWrite + ?Sized>(
        self,
        writer: &mut W,
        bits: BitCount<MAX>,
    ) -> Result<()>;
}

This trait is for integer types which can be read or written to a bit stream as a partial amount of bits.

It unifies signed and unsigned integer types by delegating reads and writes to the signed and unsigned reading and writing methods as appropriate.

Required Methods

fn read<const BITS: u32, R: BitRead + ?Sized>(reader: &mut R) -> Result<Self>
where Self: Sized,

Reads a value of ourself from the stream with the given number of bits.

Errors

Passes along any I/O error from the underlying stream. A compile-time error occurs if the given number of bits is larger than our type.

fn read_var<const MAX: u32, R>( reader: &mut R, bits: BitCount<MAX>, ) -> Result<Self>
where R: BitRead + ?Sized, Self: Sized,

Reads a value of ourself from the stream with the given number of bits.

Errors

Passes along any I/O error from the underlying stream. Also returns an error if our type is too small to hold the requested number of bits.

fn write<const BITS: u32, W: BitWrite + ?Sized>( self, writer: &mut W, ) -> Result<()>

Writes ourself to the stream using the given const number of bits.

Errors

Passes along any I/O error from the underlying stream. Returns an error if our value is too large to fit the given number of bits. A compile-time error occurs if the given number of bits is larger than our type.

fn write_var<const MAX: u32, W: BitWrite + ?Sized>( self, writer: &mut W, bits: BitCount<MAX>, ) -> Result<()>

Writes ourself to the stream using the given number of bits.

Errors

Passes along any I/O error from the underlying stream. Returns an error if our value is too small to hold the given number of bits. Returns an error if our value is too large to fit the given number of bits.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.