bitstream_io

Trait Numeric

pub trait Numeric:
    Primitive
    + Sized
    + Copy
    + Default
    + Debug
    + PartialOrd
    + Shl<u32, Output = Self>
    + ShlAssign<u32>
    + Shr<u32, Output = Self>
    + ShrAssign<u32>
    + Rem<Self, Output = Self>
    + RemAssign<Self>
    + BitOrAssign<Self>
    + BitXor<Self, Output = Self>
    + Not<Output = Self>
    + Sub<Self, Output = Self> {
    const BITS_SIZE: u32;
    const ONE: Self;

    // Required methods
    fn is_zero(self) -> bool;
    fn from_u8(u: u8) -> Self;
    fn to_u8(self) -> u8;
    fn count_ones(self) -> u32;
    fn leading_zeros(self) -> u32;
    fn trailing_zeros(self) -> u32;
    fn unsigned_value(self) -> UnsignedValue;
}

This trait extends many common integer types (both unsigned and signed) with a few trivial methods so that they can be used with the bitstream handling traits.

Required Associated Constants

const BITS_SIZE: u32

Size of type in bits

const ONE: Self

The value of 1 in this type

Required Methods

fn is_zero(self) -> bool

Returns true if this value is 0, in its type

fn from_u8(u: u8) -> Self

Returns a u8 value in this type

fn to_u8(self) -> u8

Assuming 0 <= value < 256, returns this value as a u8 type

fn count_ones(self) -> u32

Counts the number of 1 bits

fn leading_zeros(self) -> u32

Counts the number of leading zeros

fn trailing_zeros(self) -> u32

Counts the number of trailing zeros

fn unsigned_value(self) -> UnsignedValue

Convert to a generic unsigned write value for stream recording purposes

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.