bitstream_io

Module write

Traits and implementations for writing bits to a stream.

Structs

BitCounterDeprecated

For counting the number of bits written but generating no output.

BitRecorder

For recording writes in order to play them back on another writer

BitWriter

For writing bit values to an underlying stream in a given endianness.

BitsWritten

For counting the number of bits written but generating no output.

ByteWriter

For writing aligned bytes to a stream of bytes in a given endianness.

Overflowed

An error returned if performing math operations would overflow

Traits

BitWrite

A trait for anything that can write a variable number of potentially un-aligned values to an output stream

BitWrite2

A compatibility trait for older code implementing BitWrite

ByteWrite

A trait for anything that can write aligned values to an output stream

Counter

A common trait for integer types for performing math operations which may check for overflow.

ToBitStream

Implemented by complex types that don’t require any additional context to build themselves to a writer

ToBitStreamUsing

Implemented by complex types that consume additional context to build themselves to a writer

ToBitStreamWith

Implemented by complex types that require additional context to build themselves to a writer

ToByteStream

Implemented by complex types that don’t require any additional context to build themselves to a writer

ToByteStreamUsing

Implemented by complex types that consume additional context to build themselves to a writer

ToByteStreamWith

Implemented by complex types that require additional context to build themselves to a writer