bitstream_io::write

Struct BitRecorder

pub struct BitRecorder<N, E: Endianness> { /* private fields */ }

For recording writes in order to play them back on another writer

Example

use std::io::Write;
use bitstream_io::{BigEndian, BitWriter, BitWrite, BitRecorder};
let mut recorder: BitRecorder<u32, BigEndian> = BitRecorder::new();
recorder.write_var(1, 0b1u8).unwrap();
recorder.write_var(2, 0b01u8).unwrap();
recorder.write_var(5, 0b10111u8).unwrap();
assert_eq!(recorder.written(), 8);
let mut writer = BitWriter::endian(Vec::new(), BigEndian);
recorder.playback(&mut writer);
assert_eq!(writer.into_writer(), [0b10110111]);

Implementations

impl<N: Counter, E: Endianness> BitRecorder<N, E>

pub fn new() -> Self

Creates new recorder

pub fn with_capacity(bytes: usize) -> Self

Creates new recorder sized for the given number of bytes

pub fn endian(endian: E) -> Self

Creates new recorder with the given endianness

pub fn written(&self) -> N

Returns number of bits written

Panics

Panics if the number of bits written is larger than the maximum supported by the counter type. Use BitRecorder::written_checked for a non-panicking alternative.

pub fn written_checked(&self) -> Result<N, Overflowed>

Returns number of bits written

Errors

Returns an error if the number of bits written overflows our counter type.

pub fn playback<W: BitWrite>(&self, writer: &mut W) -> Result<()>

Plays recorded writes to the given writer

pub fn clear(&mut self)

Clears recorder, removing all values

impl<N: PartialOrd + Counter + Copy, E: Endianness> BitRecorder<N, E>

pub fn best<F>( self, candidate: &mut Self, f: impl FnOnce(&mut Self) -> Result<(), F>, ) -> Result<Self, F>

Returns shortest option between ourself and candidate

Executes fallible closure on emptied candidate recorder, compares the lengths of ourself and the candidate, and returns the shorter of the two.

If the new candidate is shorter, we swap ourself and the candidate so any recorder capacity can be reused.

Example

use bitstream_io::{BitRecorder, BitWrite, BigEndian};

let mut best = BitRecorder::<u8, BigEndian>::new();
let mut candidate = BitRecorder::new();

// write an 8 bit value to our initial candidate
best.write::<8, u8>(0);
assert_eq!(best.written(), 8);

// try another candidate which writes 4 bits
best = best.best(&mut candidate, |w| {
    w.write::<4, u8>(0)
}).unwrap();

// which becomes our new best candidate
assert_eq!(best.written(), 4);

// finally, try a not-so-best candidate
// which writes 10 bits
best = best.best(&mut candidate, |w| {
    w.write::<10, u16>(0)
}).unwrap();

// so our best candidate remains 4 bits
assert_eq!(best.written(), 4);