Struct bitstream_io::write::BitWriterBE

pub struct BitWriterBE<'a> { /* fields omitted */ }

A wrapper for writing values to a big-endian stream.

Methods

impl<'a> BitWriterBE<'a>

fn new(writer: &mut Write) -> BitWriterBE

Wraps a big-endian writer around a Write reference.

Because this is liable to make many small writes in the course of normal operation, a BufWrite is preferable for better performance.