Struct bitstream_io::write::BitWriterLE

pub struct BitWriterLE<'a> { /* fields omitted */ }

A wrapper for writing values to a little-endian stream.

Methods

impl<'a> BitWriterLE<'a>

fn new(writer: &mut Write) -> BitWriterLE

Wraps a little-endian writer around a Write reference.

Because this is liable to make many small writes in the course of normal operation, a BufWrite is preferable for better performance.