bitstream_io::write

Struct ByteWriter

pub struct ByteWriter<W: Write, E: Endianness> { /* private fields */ }

For writing aligned bytes to a stream of bytes in a given endianness.

This only writes aligned values and maintains no internal state.

Implementations

impl<W: Write, E: Endianness> ByteWriter<W, E>

pub fn new(writer: W) -> ByteWriter<W, E>

Wraps a ByteWriter around something that implements Write

pub fn endian(writer: W, _endian: E) -> ByteWriter<W, E>

Wraps a BitWriter around something that implements Write with the given endianness.

pub fn into_writer(self) -> W

Unwraps internal writer and disposes of ByteWriter. Any unwritten partial bits are discarded.

pub fn writer(&mut self) -> &mut W

Provides mutable reference to internal writer.

pub fn into_bitwriter(self) -> BitWriter<W, E>

Converts ByteWriter to BitWriter in the same endianness.

pub fn bitwriter(&mut self) -> BitWriter<&mut W, E>

Provides temporary BitWriter in the same endianness.

Warning

Any unwritten bits left over when BitWriter is dropped are lost.