bitstream_io::write

Trait HuffmanWrite

pub trait HuffmanWrite<E: Endianness> {
    // Required method
    fn write_huffman<T>(
        &mut self,
        tree: &WriteHuffmanTree<E, T>,
        symbol: T,
    ) -> Result<()>
       where T: Ord + Copy;
}

A trait for anything that can write Huffman codes of a given endianness to an output stream

Required Methods

fn write_huffman<T>( &mut self, tree: &WriteHuffmanTree<E, T>, symbol: T, ) -> Result<()>
where T: Ord + Copy,

Writes Huffman code for the given symbol to the stream.

Errors

Passes along any I/O error from the underlying stream.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.