bitstream_io

Module huffman

Traits and implementations for reading or writing Huffman codes from or to a stream.

Structs

WriteHuffmanTree

A compiled Huffman tree for use with the write_huffman method. Returned by compiled_write_tree.

Enums

FinalHuffmanTree 🔒

HuffmanTreeError

An error type during Huffman tree compilation.

ReadHuffmanTree

A compiled Huffman tree element for use with the read_huffman method. Returned by compile_read_tree.

WipHuffmanTree 🔒

Functions

compile_queue 🔒

compile_read_tree

Given a vector of symbol/code pairs, compiles a Huffman tree for reading.

compile_write_tree

Given a vector of symbol/code pairs, compiles a Huffman tree for writing.