bitstream_io::write

Struct BitRecorder

pub struct BitRecorder<N, E: Endianness> {
    counter: BitCounter<N, E>,
    records: Vec<WriteRecord>,
}

For recording writes in order to play them back on another writer

Example

use std::io::Write;
use bitstream_io::{BigEndian, BitWriter, BitWrite, BitRecorder};
let mut recorder: BitRecorder<u32, BigEndian> = BitRecorder::new();
recorder.write(1, 0b1).unwrap();
recorder.write(2, 0b01).unwrap();
recorder.write(5, 0b10111).unwrap();
assert_eq!(recorder.written(), 8);
let mut writer = BitWriter::endian(Vec::new(), BigEndian);
recorder.playback(&mut writer);
assert_eq!(writer.into_writer(), [0b10110111]);

Fields
counter: BitCounter<N, E>records: Vec<WriteRecord>
Implementations

impl<N: Default + Copy, E: Endianness> BitRecorder<N, E>

pub fn new() -> Self

Creates new recorder

pub fn with_capacity(writes: usize) -> Self

Creates new recorder sized for the given number of writes

pub fn endian(_endian: E) -> Self

Creates new recorder with the given endianness

pub fn written(&self) -> N

Returns number of bits written

pub fn playback<W: BitWrite>(&self, writer: &mut W) -> Result<()>

Plays recorded writes to the given writer