bitstream_io::write

Struct BitWriter

pub struct BitWriter<W: Write, E: Endianness> {
    writer: W,
    bitqueue: BitQueue<E, u8>,
}

For writing bit values to an underlying stream in a given endianness.

Because this only writes whole bytes to the underlying stream, it is important that output is byte-aligned before the bitstream writer’s lifetime ends. Partial bytes will be lost if the writer is disposed of before they can be written.

Fields
writer: Wbitqueue: BitQueue<E, u8>
Implementations

impl<W: Write, E: Endianness> BitWriter<W, E>

pub fn new(writer: W) -> BitWriter<W, E>

Wraps a BitWriter around something that implements Write

pub fn endian(writer: W, _endian: E) -> BitWriter<W, E>

Wraps a BitWriter around something that implements Write with the given endianness.

pub fn into_writer(self) -> W

Unwraps internal writer and disposes of BitWriter.

Warning

Any unwritten partial bits are discarded.

pub fn writer(&mut self) -> Option<&mut W>

If stream is byte-aligned, provides mutable reference to internal writer. Otherwise returns None

pub fn into_bytewriter(self) -> ByteWriter<W, E>

Converts BitWriter to ByteWriter in the same endianness.

Warning

Any written partial bits are discarded.

pub fn bytewriter(&mut self) -> Option<ByteWriter<&mut W, E>>

If stream is byte-aligned, provides temporary ByteWriter in the same endianness. Otherwise returns None

Warning

Any unwritten bits left over when ByteWriter is dropped are lost.

pub fn into_unwritten(self) -> (u32, u8)

Consumes writer and returns any un-written partial byte as a (bits, value) tuple.

Examples

use std::io::Write;
use bitstream_io::{BigEndian, BitWriter, BitWrite};
let mut data = Vec::new();
let (bits, value) = {
    let mut writer = BitWriter::endian(&mut data, BigEndian);
    writer.write(15, 0b1010_0101_0101_101).unwrap();
    writer.into_unwritten()
};
assert_eq!(data, [0b1010_0101]);
assert_eq!(bits, 7);
assert_eq!(value, 0b0101_101);

use std::io::Write;
use bitstream_io::{BigEndian, BitWriter, BitWrite};
let mut data = Vec::new();
let (bits, value) = {
    let mut writer = BitWriter::endian(&mut data, BigEndian);
    writer.write(8, 0b1010_0101).unwrap();
    writer.into_unwritten()
};
assert_eq!(data, [0b1010_0101]);
assert_eq!(bits, 0);
assert_eq!(value, 0);

pub fn flush(&mut self) -> Result<()>

Flushes output stream to disk, if necessary. Any partial bytes are not flushed.

Errors

Passes along any errors from the underlying stream.