Trait bitstream_io::write::HuffmanWrite

pub trait HuffmanWrite<E: Endianness> {
    // Required method
    fn write_huffman<T>(
        &mut self,
        tree: &WriteHuffmanTree<E, T>,
        symbol: T,
    ) -> Result<()>
       where T: Ord + Copy;
}

A trait for anything that can write Huffman codes of a given endianness to an output stream

Required Methods

fn write_huffman<T>( &mut self, tree: &WriteHuffmanTree<E, T>, symbol: T, ) -> Result<()>
where T: Ord + Copy,

Writes Huffman code for the given symbol to the stream.

Errors

Passes along any I/O error from the underlying stream.

Object Safety

This trait is not object safe.